rename buttones`s module
