+++ src/js/alignButtons.js
